port 443 ssl fix

+++ srcs/requirements/wordpress/wpscript.sh
@@ -3,8 +3,11 @@
 mkdir -p /var/www/localhost/htdocs
 cd /var/www/localhost/htdocs
 chmod +x wp-cli.phar
-./wp-cli.phar core download --allow-root
-./wp-cli.phar config create --dbname=wordpress --dbuser=wpuser --dbpass=password --dbhost=mariadb --allow-root
-./wp-cli.phar core install --url=localhost --title=inception --admin_user=admin --admin_password=admin --admin_email=admin@admin.com --allow-root
+
+if [ ! -f wp-config.php ]; then
+    ./wp-cli.phar core download --allow-root
+    ./wp-cli.phar config create --dbname=wordpress --dbuser=wpuser --dbpass=password --dbhost=mariadb --allow-root
+    ./wp-cli.phar core install --url=localhost --title=inception --admin_user=admin --admin_password=admin --admin_email=admin@admin.com --allow-root
+fi
 
 php-fpm83 -F
